feat: start logging diff comparing with delta
